Refactoring: Move inner classes PlayableUtils and PlayableException out of Playable
