+++ core/src/main/java/de/danoeh/antennapod/core/util/playback/Playable.java
@@ -3,14 +3,10 @@ package de.danoeh.antennapod.core.util.playback;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.os.Parcelable;
-import androidx.preference.PreferenceManager;
-import android.util.Log;
+
 import androidx.annotation.Nullable;
 import de.danoeh.antennapod.core.feed.Chapter;
-import de.danoeh.antennapod.core.feed.FeedMedia;
 import de.danoeh.antennapod.core.feed.MediaType;
-import de.danoeh.antennapod.core.preferences.PlaybackPreferences;
-import de.danoeh.antennapod.core.storage.DBReader;
 import de.danoeh.antennapod.core.util.ShownotesProvider;
 import java.util.Date;
 import java.util.List;
@@ -183,100 +179,4 @@ public interface Playable extends Parcelable, ShownotesProvider {
     @Nullable
     String getImageLocation();
 
-    /**
-     * Provides utility methods for Playable objects.
-     */
-    class PlayableUtils {
-        private PlayableUtils(){}
-
-        private static final String TAG = "PlayableUtils";
-
-        /**
-         * Restores a playable object from a sharedPreferences file. This method might load data from the database,
-         * depending on the type of playable that was restored.
-         *
-         * @return The restored Playable object
-         */
-        @Nullable
-        public static Playable createInstanceFromPreferences(Context context) {
-            long currentlyPlayingMedia = PlaybackPreferences.getCurrentlyPlayingMediaType();
-            if (currentlyPlayingMedia != PlaybackPreferences.NO_MEDIA_PLAYING) {
-                SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context.getApplicationContext());
-                return PlayableUtils.createInstanceFromPreferences(context,
-                        (int) currentlyPlayingMedia, prefs);
-            }
-            return null;
-        }
-
-        /**
-         * Restores a playable object from a sharedPreferences file. This method might load data from the database,
-         * depending on the type of playable that was restored.
-         *
-         * @param type An integer that represents the type of the Playable object
-         *             that is restored.
-         * @param pref The SharedPreferences file from which the Playable object
-         *             is restored
-         * @return The restored Playable object
-         */
-        public static Playable createInstanceFromPreferences(Context context, int type,
-                                                             SharedPreferences pref) {
-            Playable result = null;
-            // ADD new Playable types here:
-            switch (type) {
-                case FeedMedia.PLAYABLE_TYPE_FEEDMEDIA:
-                    result = createFeedMediaInstance(pref);
-                    break;
-                case ExternalMedia.PLAYABLE_TYPE_EXTERNAL_MEDIA:
-                    result = createExternalMediaInstance(pref);
-                    break;
-            }
-            if (result == null) {
-                Log.e(TAG, "Could not restore Playable object from preferences");
-            }
-            return result;
-        }
-
-        private static Playable createFeedMediaInstance(SharedPreferences pref) {
-            Playable result = null;
-            long mediaId = pref.getLong(FeedMedia.PREF_MEDIA_ID, -1);
-            if (mediaId != -1) {
-                result =  DBReader.getFeedMedia(mediaId);
-            }
-            return result;
-        }
-
-        private static Playable createExternalMediaInstance(SharedPreferences pref) {
-            Playable result = null;
-            String source = pref.getString(ExternalMedia.PREF_SOURCE_URL, null);
-            String mediaType = pref.getString(ExternalMedia.PREF_MEDIA_TYPE, null);
-            if (source != null && mediaType != null) {
-                int position = pref.getInt(ExternalMedia.PREF_POSITION, 0);
-                long lastPlayedTime = pref.getLong(ExternalMedia.PREF_LAST_PLAYED_TIME, 0);
-                result = new ExternalMedia(source, MediaType.valueOf(mediaType),
-                        position, lastPlayedTime);
-            }
-            return result;
-        }
-    }
-
-    class PlayableException extends Exception {
-        private static final long serialVersionUID = 1L;
-
-        public PlayableException() {
-            super();
-        }
-
-        public PlayableException(String detailMessage, Throwable throwable) {
-            super(detailMessage, throwable);
-        }
-
-        public PlayableException(String detailMessage) {
-            super(detailMessage);
-        }
-
-        public PlayableException(Throwable throwable) {
-            super(throwable);
-        }
-
-    }
 }
